adding multi proc
